Make tests aware of HAVE_CONFIG_H
IMHO the idea of having a configuration file header is to avoid having to pass
define arguments to the compiler, by making tests/testlib.c be aware of the
configuration file, we can avoid having to add more checks to the list.

This was necessary to avoid having to add more arguments when building with
mingw.