Update implementation to match changes done to usbi_get_device_by_session_id
Now usbi_get_device_by_session_id increments the reference count on each call
to it, so we need to be more careful on releasing the allocated resources when
calling it, to avoid leaking memory. Method get_hotplug_device_node now has an
extra boolean argument named hotplug_poll which should be set to True only when
the function hotplug_poll is working, so if a device is all ready known in the
session it's reference count shouldn't be increased.

Conflicts:

	libusb/os/windows_usb.c